M    src/org/webdsl/dsl/transformation/desugar.str
Generate a createX page for an entity X. Instead of a page
parameter carrying the entity object, the page has a local variable
declaration that is used to create a new object.

Save and Cancel buttons don't work currently, may have to do with
the fact that the transaction is not flushed before the return statement
that retrieves the identifier of the newly created object.

M    src/org/webdsl/dsl/generation/java-page.str

Generate code for local page variables (treatment similar to parameters,
instead of loading the object it may be instantiated with an initializer
expression (typically to create a new empty object)).

M    src/org/webdsl/dsl/generation/dsl-to-seam.str

Hmm; debugging code to discover what went wront with pretty-printing;
should be turned back.

M    src/Makefile

Turned down verbosity of compilation.

svn path=/webdsls/trunk/; revision=274