Working email functionality!
Email templates can be defined in pretty much the same way as page templates. For example:
define email testEmail (to_address : Email, u : UploadItem) {
  to(to_address)
  from("WebDSL applicatie", "uploadapp@webdsl.org")
  subject("Upload notificatie")
  body() {
    div("div1") {
      text(u.upload.name)
    }
  }
}


Email actions can be used as: 
  email(testEmail(emailAddress, uploadItem));
where testEmail is the name of the email template and emailAddress is a variable of type Email. The variable uploadItem is passed along in order to 
be able to use its fields in the email template.

Of course, the SMTPHOST, SMTPPORT, SMTPUSER, SMTPPASS and SMTPSSL must be set in application.ini. This can be done via the webdsl script.


svn path=/webdsls/trunk/; revision=444