Summary: Improved DB support, error messages, less debug info, new user bug.
At first the Recommendations were only available for the MySQL database, with this version you can also use other database systems.
The exception to this is H2 in memory tables, as somehow the dictionary tables are not commited in the database after being generated yet.

Recommendations now have improved error messages, and less debug notifications.
If you want to acquire a list of recommendations based on a user that was just created the Recommendation system will no longer print a stack trace of the error. 
This is caught and processed so the user will not get any recommendations yet until the next time the recommendations are determined.



svn path=/webdsls/trunk/; revision=4833