fix issue related to generated placeholder name variables
Even though they are scoped only by 'for' elements inside a template body, when template elements are lifted, the placeholder might be moved to a different template during transformations. This means the original location does not have any AST node indicating the existence of the placeholder variable.

To resolve this problem placeholders are now desugared to a placeholdervar at the top of its direct enclosing scope (first element in template body or 'for' body). The original placeholder is transformed to an expression that retrieves the deterministic and unique identifier value from the placeholder var. This way the placeholder variable is correctly passed to lifted templates, as it remains available in the original context.

svn path=/webdsls/trunk/; revision=5974