In-memory rendered page cache for anonymous and logged in users.
After checking access control, the page rendering is retrieved from cache if available. 
The cache is filled automatically and keeps the most recently used pages in cache.
All caches are invalidated when an entity change happens, if it is a session entity change, 
only the logged in user cache is invalidated. 
The logged in user cache size is for all users combined to constrain the memory usage.

add ?nocache to a url to force page render without cache

application.ini option numcachedpages=250 is default, disable with numcachedpages=0

numcachedpages setting is used for both caches (so default would be 250 anonymous and 250 logged in pages)

todo:
Make cache invalidate when facets change.
Facet indexer needs to be tweaked, should be able to depend on entity change detection instead of timed task. 

More fine-grained cache invalidation should be possible, such as per user for session entity changes.

svn path=/webdsls/trunk/; revision=5988